+++ koin-projects/koin-androidx-viewmodel/src/main/java/org/koin/androidx/viewmodel/ViewModelParameter.kt
@@ -1,5 +1,6 @@
 package org.koin.androidx.viewmodel
 
+import android.os.Bundle
 import androidx.lifecycle.ViewModelStore
 import androidx.savedstate.SavedStateRegistryOwner
 import org.koin.core.parameter.ParametersDefinition
@@ -7,9 +8,10 @@ import org.koin.core.qualifier.Qualifier
 import kotlin.reflect.KClass
 
 class ViewModelParameter<T : Any>(
-    val clazz: KClass<T>,
-    val qualifier: Qualifier? = null,
-    val parameters: ParametersDefinition? = null,
-    val viewModelStore: ViewModelStore,
-    val stateRegistryOwner: SavedStateRegistryOwner? = null
+        val clazz: KClass<T>,
+        val qualifier: Qualifier? = null,
+        val parameters: ParametersDefinition? = null,
+        val bundle: Bundle? = null,
+        val viewModelStore: ViewModelStore,
+        val stateRegistryOwner: SavedStateRegistryOwner? = null
 )