+++ koin-projects/koin-core/src/main/kotlin/org/koin/core/qualifier/StringQualifier.kt
@@ -2,6 +2,6 @@ package org.koin.core.qualifier
 
 data class StringQualifier(override val value: QualifierValue) : Qualifier{
     override fun toString(): String {
-        return "'$value'"
+        return "q:'$value'"
     }
 }