+++ koin-projects/koin-ktor/src/main/kotlin/org/koin/ktor/ext/KoinFeature.kt
@@ -20,7 +20,7 @@ import io.ktor.application.ApplicationFeature
 import io.ktor.application.ApplicationStopping
 import io.ktor.util.AttributeKey
 import org.koin.core.KoinApplication
-import org.koin.core.context.GlobalContext
+import org.koin.core.context.KoinContextHandler
 import org.koin.core.context.startKoin
 import org.koin.dsl.KoinAppDeclaration
 
@@ -37,9 +37,9 @@ class Koin {
             get() = AttributeKey("Koin")
 
         override fun install(pipeline: Application, configure: KoinAppDeclaration): Koin {
-            startKoin(configure)
+            startKoin(appDeclaration = configure)
             pipeline.environment.monitor.subscribe(ApplicationStopping) {
-                GlobalContext.stop()
+                KoinContextHandler.stop()
             }
             return Koin()
         }