+++ koin-projects/koin-test/src/main/kotlin/org/koin/test/mock/Declare.kt
@@ -1,6 +1,6 @@
 package org.koin.test.mock
 
-import org.koin.core.context.GlobalContext
+import org.koin.core.context.KoinContextHandler
 import org.koin.core.qualifier.Qualifier
 import org.koin.dsl.ModuleDeclaration
 import org.koin.dsl.module
@@ -13,7 +13,7 @@ import org.koin.test.get
  */
 fun KoinTest.declareModule(moduleDeclaration: ModuleDeclaration) {
     val module = module(override = true, moduleDeclaration = moduleDeclaration)
-    GlobalContext.get().loadModules(listOf(module))
+    KoinContextHandler.get().loadModules(listOf(module))
 }
 
 
@@ -21,7 +21,7 @@ inline fun <reified T : Any> KoinTest.declare(
         qualifier: Qualifier? = null,
         noinline instance: () -> T
 ): T {
-    val koin = GlobalContext.get()
+    val koin = KoinContextHandler.get()
     koin.declare(instance(), qualifier, override = true)
     return get(qualifier)
 }