+++ koin-projects/koin-test/src/main/kotlin/org/koin/test/mock/DeclareMock.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2018 the original author or authors.
+ * Copyright 2017-2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,97 +16,51 @@
 package org.koin.test.mock
 
 import org.koin.core.Koin
-import org.koin.core.context.GlobalContext
-import org.koin.core.definition.BeanDefinition
-import org.koin.core.error.NoBeanDefFoundException
-import org.koin.core.logger.Level
 import org.koin.core.qualifier.Qualifier
 import org.koin.core.scope.Scope
-import org.koin.core.time.measureDurationForResult
-import org.koin.ext.getFullName
 import org.koin.test.KoinTest
 import kotlin.reflect.KClass
 
 /**
- * Declare & Create a mock in Koin container for given type
+ * Declare & Stub a mock in Koin container for given type
  *
- * @author Arnaud Giuliani
+ * @author Christian Schmitz
  */
 inline fun <reified T : Any> KoinTest.declareMock(
         qualifier: Qualifier? = null,
-        noinline stubbing: (StubFunction<T>)? = null
+        secondaryTypes: List<KClass<*>> = emptyList(),
+        stubbing: StubFunction<T> = {}
 ): T {
-    val koin = GlobalContext.get()
-    return koin.declareMock(qualifier, stubbing)
+    return getKoin().declareMock(qualifier, secondaryTypes, stubbing)
 }
 
 /**
- * Declare & Create a mock in Koin container for given type
+ * Declare & Stub a mock in Koin container for given type
  *
- * @author Arnaud Giuliani
+ * @author Christian Schmitz
  */
 inline fun <reified T : Any> Koin.declareMock(
         qualifier: Qualifier? = null,
-        noinline stubbing: (StubFunction<T>)? = null
+        secondaryTypes: List<KClass<*>> = emptyList(),
+        stubbing: StubFunction<T> = {}
 ): T {
-    return _scopeRegistry.rootScope.declareMock(qualifier, stubbing)
+    return _scopeRegistry.rootScope.declareMock(qualifier, secondaryTypes, stubbing)
 }
 
 /**
- * Declare & Create a mock in Koin container for given type and scope
+ * Declare & Stub a mock in Koin container for given type and scope
  *
- * @author Pedro Moura
+ * @author Christian Schmitz
  */
 inline fun <reified T : Any> Scope.declareMock(
         qualifier: Qualifier? = null,
-        noinline stubbing: (T.() -> Unit)? = null
+        secondaryTypes: List<KClass<*>> = emptyList(),
+        stubbing: StubFunction<T> = {}
 ): T {
-    val clazz = T::class
-    val foundDefinition: BeanDefinition<T> = getDefinition(clazz, qualifier)
-    declareMockDefinition(foundDefinition, stubbing)
-    return get(qualifier)
-}
-
-inline fun <reified T : Any> Scope.declareMockDefinition(
-        foundDefinition: BeanDefinition<T>,
-        noinline stubbing: (T.() -> Unit)?
-) {
-    val definition: BeanDefinition<T> = foundDefinition.createMockedDefinition(_koin, stubbing)
-    _scopeDefinition.save(definition, forceOverride = true)
-    _instanceRegistry.saveDefinition(definition, override = true)
-}
-
-@Suppress("UNCHECKED_CAST")
-inline fun <reified T : Any> Scope.getDefinition(
-        clazz: KClass<T>,
-        qualifier: Qualifier?
-): BeanDefinition<T> {
-    _koin._logger.info("declare mock for '${clazz.getFullName()}'")
-    return _scopeDefinition.definitions.firstOrNull { it.`is`(clazz, qualifier, _scopeDefinition) } as? BeanDefinition<T>
-            ?: throw NoBeanDefFoundException("No definition found for qualifier='$qualifier' & class='$clazz'")
-}
-
-inline fun <reified T : Any> BeanDefinition<T>.createMockedDefinition(koin: Koin, noinline stubbing: (StubFunction<T>)?): BeanDefinition<T> {
-    val copy = copy(definition = {
-        if (koin._logger.isAt(Level.DEBUG)) {
-            val (mock, time) = measureDurationForResult {
-                createMock(stubbing)
-            }
-            koin._logger.debug("[Mock] created in $time ms")
-            mock
-        } else {
-            createMock(stubbing)
-        }
-    })
-    return copy
-}
-
-inline fun <reified T> BeanDefinition<T>.createMock(noinline stubbing: (StubFunction<T>)? = null): T {
     val mock = MockProvider.makeMock<T>()
-    return stubbing?.run {
-        stubbing(mock)
-        mock
-    } ?: mock
+    declare(mock, qualifier, secondaryTypes + T::class, true)
+    mock.apply(stubbing)
+    return mock
 }
 
 typealias StubFunction<T> = T.() -> Unit